requirements.txt
json
pathlib
pillow
pypdf
pdf2image
pytesseract
uroman
ocr_tamil #Tamil OCR library, only needed for processing tamil texts
